Added deleting option for custom list

+++ app.js
@@ -90,9 +90,22 @@ else
 app.post("/delete",function(req,res)
 {
   const checkedItemId=req.body.checkbox;
-  Item.findByIdAndRemove({_id:checkedItemId},function(err){});
-  res.redirect("/");
-});
+  const listName=req.body.listName;
+  if(listName==="Today")
+  {
+    Item.findByIdAndRemove({_id:checkedItemId},function(err){});
+    res.redirect("/");
+}
+else
+{
+List.findOneAndUpdate({name:listName},{$pull:{items:{ _id: checkedItemId}}},function(err,foundList)
+{
+if(!err){
+  res.redirect("/"+listName);
+}
+})
+}
+  });
 app.get("/:customListName",function(req,res){
 const customListName=req.params.customListName;
 List.findOne({name:customListName},function(err,foundList)
